feat(search): show workspace surfaces from workspace search
