+++ extensions/cmux/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Initial Version] - {PR_MERGE_DATE}
 
 - Search and jump to a workspace in cmux
+- View surfaces for a workspace directly from Search Workspaces with `Cmd+Enter`
 - Search and focus a surface (tab) across any workspace
 - Jump to the latest notification in cmux
 - Create a new workspace and focus it automatically
